requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
wincertstore==0.2
